requirements.txt
gradio==3.18.0
imutils==0.5.4
matplotlib
numpy
pandas==1.5.3
Pillow==9.4.0
tqdm==4.64.1
opencv-python==4.7.0.68

seaborn~=0.12.2
scipy~=1.10.1